vlib:  Provide named args for directed reducers
Unlike a symmetric reducer with no initial value given (in which `a` and
`b` must be the same type), a directed reducer's output and input values
may be of incompatible types, rendering those names inappropriate to the
task.  Instead, `x` will be the current element (as in `map` and `ver`),
and `L` the successive reduced results, starting with the initial value,
for a forward reducer (i.e. starting from the left).  A reverse reducer,
starting at the right end, will use `R` instead of `L` once implemented.