ADBProtocol:  Rewrite ADBCompletion()
Nitrogen's handling of UPP's, while somewhat convenient, is more bloated
than we'd like.  In particular, there's a wrapper function for 68K which
pushes the register arguments on the stack and calls the user's routine.
Normally, it would be appropriate to write a single function declared so
that the parameters are expected in registers on 68K, but the Metrowerks
compiler only permits volatile registers to be named, so ADB's use of A2
fouls this.

Instead, we apply the maxim "When in doubt, use brute force."  Write the
68K implementation in assembly language (which is actually quite easy --
the body is just one instruction (not counting RTS)) guarded by a macro.
Call the system's ADBOp() and use mac-config's UPP macros.