vlib:  Don't check for CR along with LF
There's no good reason to allow alternative line endings, and a few good
reasons not to.  In particular, it would interact quite poorly with code
signing and other integrity checks.  Also, the scenario in which several
implementations of a test for "\nend.\n" are inconsistent is a potential
security hazard.