vx:  Define CODE and DATA constants
CODE is a string consisting of everything from the beginning of the file
until the `end.` line (excluding "end." itself but including the newline
preceding it), or the entire file if no `end.` line is present.  DATA is
a string containing everything following the `end.` line (which excludes
the trailing newline in "end.\n"), unless the line is absent -- in which
case it's not defined at all.

The original source is represented either by `CODE "end.\n" DATA`, or by
`CODE` itself.