ams-core:  Install an event poll interrupt handler
Check for errors from enter_supervisor_mode(), since the attempt is made
unconditionally.  Just because we require supervisor privileges in order
to call Enqueue() and Dequeue() doesn't mean we shouldn't allow programs
with lesser requirements (e.g. QD unit tests) to be usable in user mode.

The interrupt handler itself, of course, is only run if the installation
succeeded in the first place, so a check there would be redundant.  Note
that the handler is entered in supervisor mode, and we have to switch to
user mode to call poll_user_input(), and back again to exit the handler.