ams-core:  Rewrite Dequeue() in assembly language
Check for (qEntry == queue->qHead) early, both so we don't have to check
later, and because the most frequent user of Dequeue() will be the Event
Manager (by a long margin), and most often it will be removing the first
element.  If the queue contains merely the target element, then just one
(unconditional) branch is taken.  (As a minor tradeoff, if qEntry is not
the head, that comparison gets repeated once, as part of the loop body.)