ams-core:  Disable interrupts during queue access
Disable interrupts during access to the queue structure in Enqueue() and
Dequeue(), as documented in Inside Macintosh, Volume II.  To disable, we
enter supervisor mode, set the interrupt mask to 7, and leave supervisor
mode (preserving the new interrupt mask).  Since enter_supervisor_mode()
returns the old SR in D0, we need to put the SR elsewhere before we call
enter_supervisor_mode() again to restore the old interrupt mask.  In the
case of Dequeue(), we /also/ have the function result (returned in D0 as
well) to manage, which we keep on top of the stack until just before the
closing RTS instruction.

Note that we never check the results of calling enter_supervisor_mode().
It would be possible to branch over the SR-modifying instructions in the
event it fails, but the only conceivable purpose would be for a severely
constrained subset of Mac OS in which no queues are changed at interrupt
time -- a scenario whose questionable utility wouldn't warrant the added
complexity.  Therefore, the meta-privilege of being allowed to switch to
supervisor mode is now required (and also assumed).