ams-core:  Avoid reentering wait_for_user_input()
Guard against reentry of wait_for_user_input().  This could happen if an
event-polling interrupt occurs while user-level code is already checking
for events, e.g. via WaitNextEvent() (or even SystemTask()).  Given that
the whole point of checking asynchronously is to ensure timeliness, it's
fine to just bail out when synchronous code is doing the job.