ams-core:  Implement minimal CrsrBusy flag
This is not a comprehensive solution -- it deals with one problem, which
is asynchronous calls to update_cursor_location() while the cursor is in
flux.  The flag isn't changed asynchronously, so we don't require atomic
updates.

Any drawing operation not related to the cursor itself requires that the
cursor be hidden while the drawing occurs, so calls to hide_cursor() and
show_cursor() are frequent in application code, even if the HideCursor()
and ShowCursor() traps themselves aren't called.  Similarly, in addition
to being called by SetCursor(), the update_cursor_location() function is
also called from event-checking routines.  However, since these routines
are guarded against reentry, a cursor location update via GetNextEvent()
won't be reentered by an event-polling interrupt.

Cursor routines are documented as safe to call at interrupt time, though
a caveat applies to SetCursor():  "Changing the cursor while CrsrBusy is
true causes it to leave mouse bits, or trails, on the screen."[1]  Async
code is expected to refrain from calling SetCursor() if CrsrBusy is set.
That same condition may apply to the other cursor routines; in any case,
until any code calls them asynchronously, we have nothing to test, or to
fix.

[1] Technical Note DV520 (Device Management Overview Q&As):  "How to set
    the Macintosh cursor at interrupt time"