Genie:  Fix fd storage race in root Process ctor
The fd_map structure has contiguous storage like a vector, but stretches
to accommodate newly assigned elements automatically (a bit like a map).
The problem is that the address of fds[0] can be invalidated when fds is
resized to fit fds[1].  This happens when compiling with Apple g++ 4.0.1
for Mac OS X 10.4 / PPC.

Assign fds[2] first, so that neither fds[0] or fds[1] requires expanding
the fd_map and the subexpressions are safe in any sequence.  As a bonus,
we eliminate the possibility of multiple resizings.