ams-core:  Factor out size_error()
This is a minor behavior change:  A successful call to _ReallocateHandle
will now clear MemErr.

We don't need to pad the size before comparing it to 0x7FFFFFEF, because
padding only affects the bottom two bits, and the other value is already
at the maximum in those bits, so padding doesn't change whether the size
compares greater.