ams-core:  Add MemErr-setting New/DisposePtr traps
The alloc and dealloc callback functions are installed in the trap table
slots for NewPtr() and DisposePtr().  They're enough to implement free()
and malloc(), but we can spare them Mac-OS-specific details like setting
MemErr, if we override them with our own routines.

This is a minor behavior change:  DisposePtr() now clears MemErr.