[NEW] reader:  Open a file, send the fd via stdout
In accordance with the principle of least privilege, we want to minimize
the work we do with root access.  The `reader` utility is intended to be
used in a `sudo` context, with output redirected to a Unix-domain socket
over which the fd of the opened file will be passed.  The fd may then be
used by the process receiving it, without requiring any root privileges.