ams-core:  Fix crash when cursor is at lower right
The crash occurs when the bottom of the cursor is at or below the bottom
of the screen, and the cursor's right edge is exactly at the edge of the
screen (considering the entire 16x16 grid as part of the cursor, without
regard to whether the image fills it).  The plot_cursor() routine guards
against this, but save_bits_under_cursor() didn't.