Pedestal:  Initialize memory /before/ the Toolbox
Think about it -- InitWindows(), at least, has to create WMgrPort, which
involves a bunch of memory allocations (the GrafPort itself plus several
regions).  Clearly, any memory tuning should occur first.

This bug is an unfortunate consequence of a sort of urban legend -- that
the first thing a Mac application does is to initialize the Toolbox.  It
seems like it's common knowledge, and yet it's inaccurate.