Genie, relix-kernel:  Switch to the new scheduler
Install a scheduler that maintains a wait queue and a run queue of tasks
in need of scheduling.  Previously, we relied on the system to implement
round-robin scheduling, which it more or less does, and in particular we
assumed that in between runs of the main thread, busy user threads would
get a chance to run so as to at least set a global flag to indicate that
they were, in fact, busy, and so the application shouldn't sleep for too
long in WaitNextEvent().  For some undetermined reason, this didn't work
with the custom threads we use in System 6 (which does't support Apple's
Thread Manager extension), even though it seems to work well enough with
Thread Manager.  However, the whole business was a hack in any case, and
we needn't deliberate over exactly which circumstances in which the hack
does or doesn't have the desired effect when we can uproot it wholesale.