{display,interact}-linux:  Add no-op -x for compat
The FORGE-based variants of display and interact accept -x to indicate a
magnification level, which determines the window size.  A framebuffer is
implicitly of a fixed size (i.e. that of the screen itself), and the GPU
(if one is present) will happily and efficiently enact the magnification
to fullscreen size without being asked (and in fact, regardless of being
asked otherwise).  Allow the -x switch for compatibility, but ignore it.