A-line:  Just _exit() on ToolServer user break
The problem with throwing an exception here is that main() will catch it
and call exit(), which invokes destructors for static storage -- and our
task graph may depend on objects that have already been destroyed.  (The
symptom is that dlmalloc's free() calls abort() under certain (or rather
uncertain) circumstances.  An upcoming change that eliminates the reaper
thread makes it 100% reproducible.)