Genie:  Term-notify child processes before release
[Written from a perspective in which upcoming changes had been applied.]

The notification of a terminating parent's child processes can result in
Release() calls on their behalf.  If the parent is released, destruction
will be deferred since it's the current thread.  If a child process gets
released, it will be destroyed immediately, after any destruction that's
pending.  If the parent's destruction was deferred, until forced by that
of a child, expect a crash.

Instead, notify the children before releasing the parent.