Genie:  Erase AE reply record before yielding
If we yield first, and another AESendBlocking() call begins before we've
resumed, it will have the same return ID and overwrite our reply storage
record (which we were done with), which we then erase, so it's not found
when the reply arrives and the thread is never resumed.  If we erase the
reply from the map before yielding, though, no collision occurs.  (We've
observed this occuring when running `A-line --cpp $foo` in Carbon-native
MacRelix on a high-end G5 box, and this patch indeed fixes the problem.)