ams-qd:  Don't draw rects/regions when pnVis < 0
We could short-circuit a number of region calculations in other routines
as well, but we expect the number of calls with the pen hidden to be few
(mostly from OpenRgn and OpenPicture) but unhidden to be many.  Speeding
up initial region construction at the cost of ongoing slower drawing (in
addition to larger object code) isn't a good trade.