vlib:  Rename purity to foldability
The concept of purity is related to foldability.  The latter term simply
means that the parsed expression can be evaluated early, which rules out
operations on undefined variables, as well as calls to impure functions.
(Constants which are known in advance can be propagated and incorporated
into further folding, but impure functions can't be called even if fully
defined.)

Pure functions are those with no side effects nor side causes.  That is,
they don't affect anything other than the result returned to the caller,
and they aren't affected by anything besides the caller's arguments (and
their own definition, of course).  In particular, pure functions /allow/
mutating operations, but on local variables only -- closure over symbols
from outer lexical scopes is impure, as is calling impure functions.