ams-ui:  Override SysBeep to flash menu bar
Writing BEL to stdout is fine for command-line-based programs, but there
are two issues when running GUI applications.  First, in OS X's Terminal
it's considered worthy of a Dock notification, which is beyond annoying.
On the other hand, if xv68k's stdout is redirected elsewhere (as exhibit
does), the beep gets sent to the interactor -- effectively squelching it
(at least until it fills up the socket buffer and everything grinds to a
halt).

When the Menu Manager is available, override SysBeep() to flash the menu
bar (which is what Mac OS does when the volume is zero, or audio support
doesn't exist).