v68k-mac:  Add native and emulated 'ticking' flags
Set the flags when Ticks is read, as a clue that an application could be
in a loop that does nothing but wait for Ticks to advance.  This enables
us to mitigate the disastrous performance penalties that would otherwise
apply.  The emulated flag is also a counter permitting Button() calls to
proceed without blocking.