xv68k:  Make sure events are polled in Ticks loops
If an application might be in a loop waiting for Ticks to advance, check
for events -- and reset the flag in question (which the application will
set again if it's looping as described).

Normally, a tight loop would result in event polling anyway every 65,536
instructions.  But a subsequent change will make repeated reads of Ticks
sleep until it advances, potentially running as few as thousands or even
hundreds of instructions per second.  This change makes polling not rely
on advancing the instruction count (if a Ticks-reading loop is running).