vx:  Add utime() function for microsecond clock
This has nothing to do with the obsolete POSIX function for implementing
the `touch` command: utime() is to time() as usleep() is to sleep().  We
could have named it "gettimeofday", but since its would-be namesake uses
a timeval structure for the result, I'd expect a V counterpart to return
a fractional seconds value (which I don't wish to implement instead of a
simple microseconds value).  Also, it's verbose.