vlib:  Don't overlap for-loop variable lifetimes
Previously, the prior value in the loop iterator variable wasn't getting
destroyed until the next assignment, by which point the receiver had run
and produced the next value already.

This isn't an issue for non-receiver for-loops, since in such a case the
value stored in the loop iteration variable already existed prior to the
beginning (and will still exist after the end) of the loop.