vx:  Avoid tight reference cycles in threads
V's garbage collector can deal with "loose" reference cycles, i.e. those
involving symbols.  But it can't deal with a reference cycle in which no
symbol occurs.  To prevent a thread handle from referring to itself with
no symbol in the loop, place an undefined constant in the result storage
slot in the thread handle, and store the actual result value there after
the thread function returns.