Genie:  Put the magic Users dir on the app volume
Be consistent with /sys/mac/user/home, which is a link to "Users" on the
same volume as MacRelix (if it exists at all).  Otherwise, reconstructed
symlink paths for /sys/mac/user/home and /sys/app/{dir,exe} pass through
the volume's "Users" directory and top out at /sys/mac/vol/list/N, which
is disconcerting and unwieldy.