vx:  Add thread function body to tracker root set
While the thread is running, guard its function body (and, by extension,
anything it references).  This appears to prevent expired objects during
a thread race condition.

The included test case doesn't always fail without this patch (except on
a single-core ARM computer running a Debian derivative).  Though it does
fail most of the time on Mac OS X and occasionally on Linux/x86_64.  But
with this patch, the test passes every time.