v68k-mac:  Restore regs in Toolbox dispatch path
Previously we were restoring only A0, and otherwise preserving D0 and A1
(by not using them).  Preserve D1 and D2 as well when invoking a Toolbox
trap.

We require this because the Resource Manager is documented as preserving
certain volatile registers (D1, D2, and A1 in all cases, and all of them
for LoadResource()).  Since the trap dispatcher chains to Toolbox traps,
rather than calling them directly (as it does for OS traps), it's unable
to preserve for the caller any registers clobbered by the trap function,
so it's up to the Resource Manager's trap handlers to preserve registers
on their own.  However, the trap dispatcher mustn't clobber them itself.