ams-qd:  Add {Open,Close,Kill}Poly trap handlers
Inside Macintosh, Volume I states that the Polygon structure stores each
added point as a delta from the initial point, for maximum efficiency of
OffsetPoly().  However, the actual storage format of secondary points is
undocumented (so applications shouldn't be depending on it), and until I
see evidence that translation of many-vertexed polygons is a bottleneck,
I don't see the point of optimizing it.  (OffsetPoly() can't be an issue
at this point since it's not even implemented yet.)