ams-qd:  Add Polygon-filling trap handlers
If FramePoly() were defined as framing the interior of a polygon the way
FrameRgn() frames a region, I would have included it along with the four
other drawing operations.  But Inside Macintosh, Volume I is quite clear
that each edge in a polygon is framed by a line hanging below and to the
right of the geometric line segment between the two points.  So a region
of the polygon wouldn't be used at all -- it's essentially a playback of
the recorded MoveTo() and LineTo() calls.  FramePoly() also modifies the
open region or open polygon, if present.  We need samples of these three
uses in order to verify an implementation.

The algorithm which derives a region from a polygon effectively computes
the slope of each polygon edge once per scanline, which is a performance
bottleneck (and a problem on slower devices like ARM-based computers) --
it could be improved.