xv68k:  Pass unimplemented handlers the trap word
If the trap handler address matches that of _Unimplemented, set then set
D1 to the trap word (in the native trap dispatcher only).

Previously, we were always passing the trap word in D1.  However, Inside
Macintosh documents the Resource Manager routines as preserving volatile
registers.  Since the trap dispatcher doesn't regain control afterwards,
it has to restore registers /before/ calling a Toolbox trap (unlike with
OS traps).  Fortunately, Toolbox traps don't need to know the trap word:
The only flag is auto-pop, which is purely the domain of the dispatcher,
not the trap handler.  Implementing this change broke Pack4 et al, since
they were sharing a handler which relied on the trap word in D1 to route
the call to the appropriate package routine, but this was fixed by using
separate entry points.  However, the handler called for an unimplemented
trap is necessarily shared by all unimplemented traps -- comparing one's
trap handler address to that of _Unimplemented is how programs determine
if a trap is implemented in the first place.  It would be very useful if
this handler were to report the unimplemented trap number, and therefore
we make an exception and pass the trap word in this case.

Stashing the trap word in low memory (e.g. ToolScratch) isn't a workable
alternative, because an asynchronous Toolbox trap call could clobber it.