ams-qd:  Reuse temporary dst port in CopyBits
CopyBits() is a simple wrapper around the StdBits() bottleneck function,
which takes nearly the same arguments but omits dstBits, instead relying
on qd.thePort->portBits.  If qd.thePort->portBits happens to be dstBits,
then we're good to go; if not, we need to temporarily set a scratch port
as thePort and set its portBits to dstBits.

Instead of allocating a scratch port each time, do it only once and then
reuse it when needed.