ams-ui:  Just check the update region
Don't intersect the update region with the visRgn to check if the window
/really/ needs updating.  Most of the time, updates will be visible, and
only rarely offscreen or behind another window.  It's likely the cost of
checking more thoroughly exceeds the cost of redrawing obscured content.

Also, it turns out visRgn is really supposed to be in local coordinates.
So once we've fixed that, the extra check would additionally require two
OffsetRgn() calls: one to translate local to global, and one to go back.