ams-qd, ams-ui:  Store visRgn in local coordinates
Inside Macintosh, Volume I doesn't address a visRgn's coordinate system.
It merely says, "The default visRgn is set to the portRect."  Given that
the portRect is in local coordinates, this seems to imply that visRgn is
as well.  This is confirmed by Lode Runner, which passes the destination
window's visRgn as the maskRgn to CopyBits() (which is superfluous since
CopyBits() already clips to the visRgn when copying to thePort's bitmap,
as in this case).  With a maskRgn erroneously in global coordinates, the
top and left of the window will remain blank.