ams-qd:  Hide the cursor when copying screen bits
This fixes a cute little bug that can be seen in Amazing:  Launch it and
immediately choose New from the menu.  Solve the maze (by tracing a path
with the cursor from start to finish) to display the victory image, then
backtrack to remove it.  There'll be a ghost image of the cursor exactly
where it was when the maze was solved, because while we took measures to
temporarily erase the cursor while drawing or blitting to the screen, we
didn't do so for copying bits /from/ the screen, and those bits are then
copied back to the screen (rather than redrawing the maze the hard way).