ams-qd:  Default to patCopy, save/restore PenState
Prologue from The Fool's Errand has one picture which consists solely of
short line segments.  It expects the pen mode to be patCopy, even though
it doesn't include opcode $08 (PnMode).  (It sets the pen pattern to all
white pixels, so none of patOr, patXor, or patBic would have an effect.)