ams-seg:  Store a Finder info handle in two places
The Fool's Errand depends on this.  In version 2.0, without this change,
it dereferences the null handle (successfully reading zero as a pointer,
since address $0000 has been mapped in order to avoid crashes elsewhere)
and tries to read 2 bytes from $0002.  Inside Macintosh doesn't document
what lives in the 32 bytes between the address in A5 and the jump table,
but the Segment Loader chapter discussing the jump table also spells out
a "Finder information" handle, stored in the global AppParmHandle, whose
name suggests a relation to the "application parameters" labeled in this
diagram:

<http://mirror.informatimago.com/next/developer.apple.com/documentation/mac/Processes/Processes-138.html#HEADING138-0>

    MOVEA.L  (16,A5),A0
    MOVEA.L  (A0),A0
    MOVE.W   (2,A0),D6

Sure enough, the book /How to Write Macintosh Software/ by Scott Knaster
confirms this in the segment "Play by Play: An Application Is Launched".
Step 21 (page 376) says "move the Finder information handle to its place
at $10(A5)".

Version 3.0 of The Fool's Errand checks for a null handle, but its error
recovery code path (which presumably was never tested, since real Mac OS
never provides the failure mode) ends up blindly dereferencing A0, which
in this case has a leftover odd value, yielding an Address Error.  (This
was the first of the two I ran into and a /lot/ harder to debug than the
other would have been.)

    MOVE.L   (16,A5),D6
    BEQ.S    *+26
    MOVEA.L  D6,A0
    ...
    MOVE.W   (A0),D0

Anyway, create a zero-element Finder information handle, storing it both
in AppParmHandle (as documented in Inside Macintosh, Volume II) and also
in the application parameters area above A5.