ams-qd:  Implement Picture opcode $90, BitsRect
Previous testing showed that, given a PICT with a PackBitsRect operation
that has rowBytes < 8 but is otherwise well-formed, Mac OS will crash --
apparently it's disregarding the actual opcode and depending entirely on
rowBytes.  Perhaps it does so to share the majority of the code path, as
we're now doing.

"The Sun's Map has been corrupted by a powerful enchantment."  You think
it's hard now?  Try it /without/ this patch...