ams-fs:  Save trap handler addresses set by ams-io
Hold onto these so we can call them from our own handlers.  We'd already
been doing this for Open; now additionally cover Close, Read, and Write.
These four are the only traps shared between the Device Manager and File
Manager.  For the latter three, we'll forward to ams-io if the refNum is
negative (indicating a driver rather than a file).