Revert "kill-systemstats:  Run under sudo"
Turns out sudo, like killall itself, relies on the real user ID, not the
effective user ID, so sudo still asks for a password when running setuid
root, just like killall only tries to match the calling user's processes
(despite having the privileges to kill any process).  The solution is to
call setuid() to set the real user ID, which obviates the need for sudo.

This reverts commit 7e83fe8b86cbd98dc87000ddec6c62bb3ee9d00f.