ams-core:  Initialize the cursor automatically
It seems that /some/ applications don't call InitCursor() as part of the
Mac Toolbox initialization sequence.[1]  So that means that /we/ have to
initialize the cursor, instead of depending on the application to do it.

[1] Here I'm referring to Lode Runner in very much particular.

P.S. Curiously, this patch exposed a thornier cursor-drawing bug in Lode
Runner, whose nature still eludes me, requiring the patch to be deferred
until it was resolved.  Fortunately, the issue is nonexistent when using
the fork-based GetResource() implementation.