ams-qd:  Check for equal bitmaps in CopyBits()
CopyBits() is a simple wrapper around the StdBits() bottleneck function,
which takes nearly the same arguments but omits dstBits, instead relying
on qd.thePort->portBits.  If qd.thePort->portBits happens to be dstBits,
then we're good to go; if not, we need to temporarily set a scratch port
as thePort and set its portBits to dstBits.

Instead of requiring identical pointer addresses, allow a looser member-
wise comparison.