ams-core:  Use low mem vs. QDGlobals in SetMouse()
Bounds-check the mouse location using the low memory global CrsrPin, not
qd.screenBits.bounds.  QuickDraw globals are initialized when InitGraf()
is called by the application, which is after all the modules are loaded.
This leaves ample time for the event polling interrupt (which runs every
65536 instructions) to receive a mouse movement.

This fixes the error "Illegal memory access: 2-byte read at $ffffff92 in
User Data Space".