[NEW] writer:  Open a file, send the fd via stdout
In accordance with the principle of least privilege, we want to minimize
the work we do with root access.  The `writer` utility is intended to be
used in a `sudo` context, with output redirected to a Unix-domain socket
over which the fd of the opened file will be passed.  The fd may then be
used by the process receiving it, without requiring any root privileges.

This program is identical to its older sibling called `reader`[1] except
for opening the target file O_WRONLY instead of O_RDONLY.

[1] See 4af2515afb83a8c292e4d0d345e7b73585d507dd.