display-linux:  Launch kdmode as a coprocess
The overarching problem here is that we want to be able to run graphical
programs (such as AMS) from the command line.  The framebuffer's console
must initially be in text mode to enter the command, but then we want to
switch to graphics mode to disable the screensaver, cursor blinking, and
any other distractions.

However, switching console modes requires root.  Running the entire user
application is out of the question.  At the moment, we rely on specific,
highly focused utilites that are setuid-root.  While a vast improvement,
we can do better.  The new approach uses sudo instead of a setuid bit to
elevate privileges.  Since asking again for a mode-changing password now
that we're in graphics mode is again out of the question, the privileged
program has to be a coprocess that stays around until we're through with
it.

On the other hand, running sudo takes about 166ms on a Raspberry Pi, and
we have to run it twice -- once to ascertain whether a password would be
required (so we know whether to issue a prompt for it), and once more to
actually run the command.  So we continue to support setuid-root utility
programs as an option, in addition to sudo.

It's possible that the console is already in graphics mode, and we don't
need root at all.  /dev/tty0 is writable by group tty in recent kernels,
so if the user is in the tty group, then we can just open the device and
ask (but not actually switch modes -- only root can do that).  But on an
older kernel, it takes root just to open the device; in that case, we'll
silently move on.

By initiating the mode switch from display-linux itself, rather than one
of its caller ancestors, we have additional flexibility.  For example, a
mode that preserves the console text might change the bit depth /before/
switching to graphics mode -- with the text redrawn at the new depth for
free, instead of having to buffer/transcode it explicitly.  Conceivably,
we might even prefer to stay in text mode to permit simultaneous console
use, or toggle modes briefly to display messages written to the console.