Install AMS root utils in /usr/local/lib/metamage/
A program that (possibly) must run as root needs to be stored in a known
location, so the caller can pass a full path to sudo (which won't search
the caller's PATH).  Until the install prefix can be configured at build
time, we need to pick one.  Respect the standards and prefer /usr/local/
to /opt/.