plus:  Use boost::atomic refcount in 64-bit builds
Since OTAtomicAdd32() isn't available in 64-bit builds anyway, we may as
well switch to Boost.Atomic since there's nothing to lose.  Even better,
the "impossible constraint in 'asm'" error we encountered on ppc and x86
doesn't occur on x86_64!  On the other hand, ppc64 still fails -- though
at least it compiles, and presumably it could even be made to work if it
were properly linked with a Boost library defining scoped_lock routines.