pass_fd:  Appease incompatibly offended dev tools
The cmsg_fd_alloc aggregate can't continue to be a struct, because glibc
has inexplicably placed a flexible array member inside the cmsghdr type,
preventing any following members.  But it also can't be a union, because
Apple's g++ 4.0 informs us that CMSG_SPACE(sizeof (int)) is not actually
a constant integer expression!  So it can't be either one -- which means
that ultimately, it must be both.  The good news is that the conditional
test is obvious, at least -- the only broken g++ of interest is Apple's.