vlib:  Fix removal of wrong GC root in del_root()
Comparing values for identity with expr() is broken for non-pair objects
which may harbor reference cycles, like threads.  However, since a value
that's capable of referencing another must necessarily have an allocated
extent, we can use the extent pointer as a unique object ID.

The symptom was two failure modes in pair-up.vx.  Both were derived from
the same tendency of del_root() to remove a thread instead of some other
object that was the intended target, thereby destroying the thread.  The
only difference was whether the hapless thread was in fact the very same
one doing the destroying (resulting in a self-join, which is detected by
the runtime, terminating it), or some other thread, which in tossup.vx's
case led to deadlock due to the bit commitment protocol.

Until a85aa7871311e1f53c3d1af764c00323f89d8f38, this was highly visible.
The bug seems to have a better than even chance of occurring in a run of
100 tossups.  e4adb33ac99f63ad9bf72ff9a5ea7551997ed3ae added a "spoiled"
version of pair-up.vx that achieves the same effect in only ten tossups,
by adding otherwise useless GC churn.  With this change, pair-up-spoiled
can survive 100 tossups in a row (and has yet to fail at all).  If there
is even a one half chance of the bug escaping ten tossups, then the odds
of slipping past 100 tossups unseen is less than one in a thousand -- so
it's fair to conclude that this patch works as expected.

The reason a85aa7871311e1f53c3d1af764c00323f89d8f38 had to be applied in
advance is so that the corrected object identity comparison didn't yield
undefined behavior due to accessing uninitialized memory (in the objects
which didn't belong in the root set anyway) and corresponding diagnostic
warnings in valgrind.