vlib:  Don't clear cycle-free flag on unshare
This was an attempt to be conservative while splicing lists together, in
case a non-cycle-free value was spliced onto the end of a chain that was
otherwise cycle-free.  However, this was shown to be a doomed effort, as
seen in d3bf16876f1c4c1a3c9b99e2f9d8758c410b15c0.  Worse, it was causing
simple values like strings, even those with no extent, to be added as GC
roots -- which is not only bad for performance, but badly interacts with
an imminent GC bugfix.