vlib:  Take GC lock when assigning variables
This irritatingly creates another dependency loop, but it does address a
race observed between a thread assigning a symbol and another collecting
garbage.

The loop could be broken by moving gc_lock to a new unit, but that would
complicate the task of ensuring that gc_mutex is initialized before use.