[NEW] SoundDriver:  Missing API for classic Mac OS
The Sound Driver API was deprecated with extreme prejudice.  The Sound.h
header in Universal Interfaces only mentions it to announce how obsolete
it is.  But Universal Interfaces can otherwise target the older versions
of Mac OS that still have a functioning .Sound driver, so this interface
is useful to have.