ams-core:  Return qErr from Dequeue of null entry
If a caller merely calls `Dequeue( qHdr.qHead, &qHdr )` without checking
for an empty queue (which is possible due to a race condition), politely
return qErr instead of crashing (by dereferencing NULL).

It's already the case that Dequeue() might return qErr for a queue head,
since a competing call might win the race and dequeue the element first.
Since callers therefore have to check the result of Dequeue() to know if
they should proceed, we may as well spare them the burden of a preceding
null-check as well.