ams-qd:  Use fast_rshift() for misaligned blitting
This cuts NyanCat's CPU use in half if the window is moved over a pixel.
(By default, the window is placed so that the left edge is already byte-
aligned, so it has to be moved to see an effect.)  On the other hand, it
doesn't benefit Stuntcopter or Lode Runner -- which makes sense, because
they blit much narrower bands of pixels.