ams-core:  Fix premature return from Delay()
With non-negative numTicks, Delay() is supposed to return only after the
specified time has elapsed, but instead it was exiting if reactor_wait()
returned false.  The obvious symptom was highly rapid menu item blinking
in Lode Runner (which installs a VBL task).

While we're here, fix a correctness issue with casting negative numTicks
arguments:  The result of casting a negative number to uint64_t and then
multiplying will invariably overflow.  We only need to cast to uint32_t,
leaving us with around two to four billion ticks (which should be enough
for anybody).  Even if an I/O request really were pending for that long,
the synchronous call implementation calls Delay(-1, x) in a loop anyway.