ams-ui:  Don't issue non-dialog updates filterless
That is, while a call to ModalDialog() with null filterProc is in scope,
return false from CheckUpdate() for non-dialog windows, even if they are
due for an update event.  Otherwise, if an application initiates a modal
dialog with a non-dialog update pending, WaitNextEvent() will provide an
update event every time, immediately, and never actually do any waiting.