ams-ui:  Update non-frontmost dialog windows
Don't assume that a modal dialog is the only dialog visible.  If we have
an update pending for any dialog, draw it.  (This may cause problems for
applications passing a non-null filterProc, since there might be updates
for non-dialog windows.  The solution is to support filterProcs at all.)

Otherwise, Lode Runner's "Unable to save high scores" alert pegs a core.