ams-ui:  Reexpand DITL text in ParamText()
Some applications (including Gunshy) call ParamText() /after/ creating a
dialog box and expect it to take effect.  Expand the text in NewDialog()
as before, but also do so in ParamText() if the front window is a dialog
box.

Inside Macintosh claims "ParamText... replaces the special strings... in
the static text items of all ***subsequently created*** alert and dialog
boxes with the text strings you pass as parameters." (strenuous emphasis
mine)  But later on it says "If a modeless dialog box using ParamText is
onscreen and you display another dialog box... that also uses ParamText,
both boxes will be affected by the latest call to ParamText" -- which is
in conflict with the first statement.

<http://mirror.informatimago.com/next/developer.apple.com/documentation/
mac/Toolbox/Toolbox-417.html>