xv68k:  Account for PEA in patched spinloop code
We're appropriating trap A0F6 to mitigate BNE spinloops.  The patch will
replace `TST.W <ea>` with `PEA <ea>` and `BNE.S` with the A-trap.  Since
we'd rather not make the trap handler dependent on the trap dispatcher's
internal details (i.e. how much it pushes onto the stack before it calls
the trap handler), pop the watched memory address off the stack first if
we're dispatching the new trap, and load the value into A0 (though after
saving the previous value on the stack).  The patch is only applied with
the native trap dispatcher in use, so we only need to address that case.