ams-core:  Maintain CrsrRect and CrsrAddr
CrsrRect and CrsrAddr are now set at the top of paint_cursor(), but then
CrsrAddr is reassigned.  The new value set in set_Crsr_vars() is defined
by a CrsrRect that shrinks as the cursor crosses CrsrPin.bottom, but the
old value that replaces it corresponds to a cursor rect that's always 16
rows tall.