xv68k:  Fix event poll interrupt deadlock
If the event poll interrupt fired when interrupts were disabled, it just
got lost -- which wasn't really a big deal, since it would fire again in
another 65536 instructions... unless at /that/ point interrupts happened
to be disabled also, and it turns out that the application is waiting in
a loop for MBState to change, and the period of the loop happens to be a
factor of 65536, so interrupts are disabled *every* time, and low memory
globals never get updated nor events posted.  So defer to later instead.

This fixes clicking the deck in Solitaire and clicking to unhighlight in
GunShy.