ams-core:  Return void from check_timers()
We aren't currently using the result.  Also, we might want to change the
semantics in the future, so remove it now.  The present behavior returns
true if any timers remain in the queue at the end of check_timers().  An
alternative is returning true only if any timers fired and were removed.