exosnd-api, ams-io:  Add a record ID to four-tones
While in practice, an FTSoundRec modified dynamically will invariably be
the one currently playing, it's possible that a pending four-tone record
could be changed in advance.  Therefore we need to tag four-tone records
on the wire with a unique ID, for which we can just use the FTSoundRec's
address in emulated memory.